+++ packages/api/lib/index.ts
@@ -1,4 +1,3 @@
 export * from './generated/sdk';
 export * from './api-client';
-export * from './seamless-api-client';
 export * from './constants/index';
